+++ CHANGELOG.rst
@@ -20,6 +20,13 @@ The changes listed in this file are categorised as follows:
 ### Fixed
 - Fixes to generate annual and monthly gridded ensembles with unified noise and preserving more of the variance.
 
+### Changed
+
+- Now possibly to send variable length temperature scaling timeseries, fixed noise generator for wrong ordering of base data dimensions
+
+### Fixed
+
+- Variable length timeseries now works also when don't have "year" as time dimension.
 
 
 [Version 1.6.0]